code/recommender/requirements.txt
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-surprise==1.1.4
scipy==1.14.1
six==1.16.0
tzdata==2024.2
